refactor(cli): move transform utils with new variablename
